% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-stat-df.R
\name{build_stat_df}
\alias{build_stat_df}
\title{Take a vector as input and return a tibble with the min, mean, median and max of this vector}
\usage{
build_stat_df(x)
}
\arguments{
\item{x}{}
}
\value{
data frame min | mean | var | max
}
\description{
Take a vector as input and return a tibble with the min, mean, median and max of this vector
}
