% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_map_df.R
\name{compute_map_df}
\alias{compute_map_df}
\title{Compute the 2 data frame needed to plot the maps from the filtered one.}
\usage{
compute_map_df(filtered_df)
}
\arguments{
\item{filtered_df}{a filtered version of the full data}
}
\value{
A list of dataframe `heatmap` containing "Number of Intervention", "Mean Intervention Duration", "Mean Number of Units" per zip code. A list of dataframe `fire_box` containing "Number of Intervention", "Mean Intervention Duration", "Mean Number of Units" per fire box.
}
\description{
Compute the 2 data frame needed to plot the maps from the filtered one.
}
