% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-fdny.R
\name{filter_fdny}
\alias{filter_fdny}
\title{Apply the filter to the main data frame}
\usage{
filter_fdny(df, input)
}
\arguments{
\item{df}{main data frame}

\item{input}{shiny /list/ inputs}
}
\value{
subset of the param data frame filtered following the input
}
\description{
Apply the filter to the main data frame
}
