% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-fdny.R
\name{filter_fdny}
\alias{filter_fdny}
\title{Filter the data frame containing all the data using the inputs from the shiny app.}
\usage{
filter_fdny(df, input)
}
\arguments{
\item{df}{data frame containing all the data}

\item{input}{input from the shiny app}
}
\value{
A filtered version of the full data frame. A subset of the full dataframe.
}
\description{
Filter the data frame containing all the data using the inputs from the shiny app.
}
