% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-deployment-time.R
\name{get_deployment_time}
\alias{get_deployment_time}
\title{Extract the deployment time from a data frame with the same columns as tidy_incidents data.}
\usage{
get_deployment_time(df)
}
\arguments{
\item{df}{data frame from which we want to extract the deployment time}
}
\value{
data frame with a column named `col` containing all the deployment time.
}
\description{
Extract the deployment time from a data frame with the same columns as tidy_incidents data.
}
