% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-intervention-duration.R
\name{get_intervention_duration}
\alias{get_intervention_duration}
\title{Extract the intervention duration from a data frame with the same columns as tidy_incidents data.}
\usage{
get_intervention_duration(df)
}
\arguments{
\item{df}{data frame from which we want to extract the intervention duration}
}
\value{
data frame with a column named `col` containing all the intervention duration.
}
\description{
Extract the intervention duration from a data frame with the same columns as tidy_incidents data.
}
