% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-interventions-per-box.R
\name{get_inteventions_per_box}
\alias{get_inteventions_per_box}
\title{Extract the interventions per box from a data frame with the same columns as tidy_incidents data.}
\usage{
get_inteventions_per_box(df)
}
\arguments{
\item{df}{data frame from which we want to extract the interventions per box}
}
\value{
data frame with a column named `col` containing all the interventions per box.
}
\description{
Extract the interventions per box from a data frame with the same columns as tidy_incidents data.
}
