% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nb_units.R
\name{get_nb_units}
\alias{get_nb_units}
\title{Extract the number of units from a data frame with the same columns as tidy_incidents data.}
\usage{
get_nb_units(df)
}
\arguments{
\item{df}{data frame from which we want to extract the number of units}
}
\value{
data frame with a column named `col` containing all the number of units.
}
\description{
Extract the number of units from a data frame with the same columns as tidy_incidents data.
}
