% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nyc_map_firebox_full.R
\docType{data}
\name{nyc_map_firebox_full}
\alias{nyc_map_firebox_full}
\title{Data frame containing the New York fire boxposition}
\format{A data frame with 16292 rows and 3 variables:
\describe{
  \item{Name}{fire box name as a Factor}
  \item{Description}{Position description}
  \item{geometry}{Position}
}}
\source{
\url{https://opendata.cityofnewyork.us/}
}
\usage{
nyc_map_firebox_full
}
\description{
Data frame containing the New York fire boxposition
}
\keyword{datasets}
