% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firebox_layer.R
\name{plot_firebox}
\alias{plot_firebox}
\title{Plot the map containing the infos of each fire box.}
\usage{
plot_firebox(fireboxes_df)
}
\arguments{
\item{fireboxes_df}{a df containing  n | mean_interv_time | mean_nb_units for each fire_box}
}
\value{
Plot of the fire box with corresponding data
}
\description{
Plot the map containing the infos of each fire box.
}
