% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_leaflet_map.R
\name{plot_leaflet_map}
\alias{plot_leaflet_map}
\title{Extract the number of units from a data frame with the same columns as tidy_incidents data.}
\usage{
plot_leaflet_map(map_data_df, input)
}
\arguments{
\item{map_data_df}{list of dataframe `heatmap` containing n, mean_interv_time, mean_nb_units per zip code. A list of dataframe `fire_box` containing n, mean_interv_time, mean_nb_units per fire box.}

\item{input}{input from the shiny app.}
}
\value{
plot a leaflet heatmap for the map_data_df
}
\description{
Extract the number of units from a data frame with the same columns as tidy_incidents data.
}
