% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{server}
\alias{server}
\title{The Shiny app server}
\usage{
server(input, output, session)
}
\arguments{
\item{input}{input from ui shiny}

\item{output}{output for ui shiny}

\item{session}{session}
}
\value{
shiny app server
}
\description{
The Shiny app server
}
