% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-deployment-time.R,
%   R/get-intervention-duration.R, R/get-interventions-per-box.R,
%   R/get-nb-units.R, R/statistic-fdny.R
\name{get_deployment_time}
\alias{get_deployment_time}
\alias{get_intervention_duration}
\alias{get_inteventions_per_box}
\alias{get_nb_units}
\alias{statistic_fdny}
\title{Give the statistic dataframe}
\usage{
get_deployment_time(df)

get_intervention_duration(df)

get_inteventions_per_box(df)

get_nb_units(df)

statistic_fdny(df, input)
}
\arguments{
\item{df}{main data frame}

\item{input}{shiny /list/ inputs}

\item{df}{main data frame}

\item{input}{shiny /list/ inputs}

\item{df}{main data frame}

\item{input}{shiny /list/ inputs}

\item{df}{main data frame}

\item{input}{shiny /list/ inputs}

\item{df}{main data frame}

\item{input}{shiny /list/ inputs}
}
\value{
data frame filtered and data frame with corresponding statistics

data frame filtered and data frame with corresponding statistics

data frame filtered and data frame with corresponding statistics

data frame filtered and data frame with corresponding statistics

data frame filtered and data frame with corresponding statistics
}
\description{
Give the statistic dataframe

Give the statistic dataframe

Give the statistic dataframe

Give the statistic dataframe

Give the statistic dataframe
}
