% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic_fdny.R
\name{statistic_fdny}
\alias{statistic_fdny}
\title{Give the statistic dataframe}
\usage{
statistic_fdny(df, input)
}
\arguments{
\item{df}{main data frame}

\item{input}{shiny /list/ inputs}
}
\value{
data frame filtered and data frame with corresponding statistics
}
\description{
Give the statistic dataframe
}
